requirements.txt
streamlit>=1.28.0
google-cloud-bigquery>=3.11.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0 